+++ src/Widget/ExplanationField.php
@@ -15,16 +15,41 @@ class ExplanationField extends Widget
 {
     protected $strTemplate = 'be_explanation';
 
+    protected $collapsible = false;
+
     public function __construct(array $attributes = null)
     {
         parent::__construct($attributes);
         if (isset($attributes['text_callback']) && \is_array($callback = $attributes['text_callback'])) {
             $this->text = System::importStatic($callback[0])->{$callback[1]}($attributes);
         }
+        if (isset($attributes['collapsible']) && true === $attributes['collapsible']) {
+            $this->collapsible = true;
+        }
     }
 
     public function generate()
     {
-        return parent::generate();
+        return $this->parse();
     }
+
+    /**
+     * Parse the template file and return it as string
+     *
+     * @param array $arrAttributes An optional attributes array
+     *
+     * @return string The template markup
+     */
+    public function parse($arrAttributes = null)
+    {
+        if ('BE' === TL_MODE) {
+            $GLOBALS['TL_CSS']['be_explanation'] = 'bundles/heimrichhannotcontaobeexplanation/assets/contao-be-explanation-bundle.css|static';
+            if ($this->collapsible) {
+                $GLOBALS['TL_JAVASCRIPT']['be_explanation'] = 'bundles/heimrichhannotcontaobeexplanation/assets/contao-be-explanation-bundle.js|static';
+            }
+        }
+        return parent::parse($arrAttributes);
+    }
+
+
 }